+++ src/core/Info/parser/Html5Player.ts
@@ -1,6 +1,5 @@
-type Html5PlayerInfo = { url: string; body: string | null; id: string; signatureTimestamp: string };
-
-import type { YTDL_GetInfoOptions } from '@/types/Options';
+import type { YTDL_DownloadOptions, YTDL_GetInfoOptions } from '@/types/Options';
+import type { Html5PlayerCache } from '@/types';
 
 import { Platform } from '@/platforms/Platform';
 
@@ -8,7 +7,8 @@ import { Signature } from '@/core/Signature';
 import { Fetcher } from '@/core/Fetcher';
 
 import { Url } from '@/utils/Url';
-import { CURRENT_PLAYER_ID } from '@/utils/Constants';
+import { Logger } from '@/utils/Log';
+import { getDecipherFunction, getNTransformFunction } from '@/utils/Html5Player';
 
 const SHIM = Platform.getShim(),
     GITHUB_API_BASE_URL = `https://raw.githubusercontent.com/${SHIM.info.repo.user}/${SHIM.info.repo.name}/dev/data/player`,
@@ -28,55 +28,61 @@ function getPlayerId(body?: string): string | null {
     return null;
 }
 
-async function getHtml5Player(options: YTDL_GetInfoOptions): Promise<Html5PlayerInfo> {
-    const CACHE = await FileCache.get<Html5PlayerInfo>('html5Player');
+async function getPlayerFunctions(options: YTDL_GetInfoOptions, html5Player: YTDL_DownloadOptions['html5Player']): Promise<Html5PlayerCache> {
+    const CACHE = await FileCache.get<Html5PlayerCache>('html5Player');
 
-    if (CACHE && CACHE.url) {
-        return {
-            url: CACHE.url,
-            body: CACHE.body,
-            id: CACHE.id,
-            signatureTimestamp: CACHE.signatureTimestamp,
-        };
+    if (CACHE && CACHE.signatureTimestamp) {
+        return CACHE;
     }
 
+    Logger.debug('To speed up processing, html5Player and signatureTimestamp are pre-fetched and cached.');
+
     let playerId = undefined,
-        signatureTimestamp = undefined;
+        playerBody = undefined;
 
-    try {
-        const IFRAME_API_BODY = await Fetcher.request<string>(Url.getIframeApiUrl(), {
-            ...options,
-            requestOptions: {
-                headers: { 'x-browser-channel': 'stable', 'x-browser-copyright': 'Copyright 2024 Google LLC. All rights reserved.', 'x-browser-validation': 'g+9zsjnuPhmKvFM5e6eaEzcB1JY=', 'x-browser-year': '2024' },
-            },
-        });
-        playerId = getPlayerId(IFRAME_API_BODY);
-    } catch {}
-
-    if (!playerId) {
+    if (!html5Player?.useRetrievedFunctionsFromGithub) {
         try {
-            const GITHUB_PLAYER_JSON = JSON.parse(await Fetcher.request<string>(GITHUB_API_BASE_URL + '/data.json'));
-            playerId = GITHUB_PLAYER_JSON.playerId;
-            signatureTimestamp = GITHUB_PLAYER_JSON.signatureTimestamp;
+            const IFRAME_API_BODY = await Fetcher.request<string>(Url.getIframeApiUrl(), options);
+            playerId = getPlayerId(IFRAME_API_BODY);
         } catch {}
     }
 
-    if (!playerId) {
-        playerId = CURRENT_PLAYER_ID;
+    if (html5Player?.useRetrievedFunctionsFromGithub || !playerId) {
+        const GITHUB_PLAYER_DATA = await Fetcher.request<string>(GITHUB_API_BASE_URL + '/data.json');
+        FileCache.set('html5Player', GITHUB_PLAYER_DATA);
+        return JSON.parse(GITHUB_PLAYER_DATA);
     }
 
-    const PLAYER_URL = Url.getPlayerJsUrl(playerId),
-        HTML5_PLAYER_BODY = (PLAYER_URL ? await Fetcher.request<string>(PLAYER_URL, options) : '') || (await Fetcher.request<string>(GITHUB_API_BASE_URL + '/base.js')),
-        DATA = {
-            url: PLAYER_URL,
-            body: HTML5_PLAYER_BODY || null,
-            id: playerId,
-            signatureTimestamp: signatureTimestamp || (PLAYER_URL ? Signature.getSignatureTimestamp(HTML5_PLAYER_BODY) || '' : ''),
-        };
+    const PLAYER_URL = Url.getPlayerJsUrl(playerId);
+    if (PLAYER_URL) {
+        try {
+            playerBody = await Fetcher.request<string>(PLAYER_URL, options);
+        } catch {}
+    }
+
+    if (!playerBody) {
+        try {
+            playerBody = await Fetcher.request<string>(GITHUB_API_BASE_URL + '/base.js');
+        } catch {}
+    }
+
+    if (!playerBody) {
+        throw new Error('Failed to retrieve player body.');
+    }
+
+    const DATA: Html5PlayerCache = {
+        id: playerId,
+        body: playerBody,
+        signatureTimestamp: PLAYER_URL ? Signature.getSignatureTimestamp(playerBody) || '' : '',
+        functions: {
+            decipher: getDecipherFunction(playerId, playerBody),
+            nTransform: getNTransformFunction(playerId, playerBody),
+        },
+    };
 
     FileCache.set('html5Player', JSON.stringify(DATA));
 
     return DATA;
 }
 
-export { getHtml5Player };
+export { getPlayerFunctions };