+++ src/dependence/Jinter/nodes/ArrowFunctionExpression.ts
@@ -3,28 +3,28 @@ import { namedFunction } from '../utils/index.js';
 import BaseJSNode from './BaseJSNode.js';
 
 export default class ArrowFunctionExpression extends BaseJSNode<ESTree.ArrowFunctionExpression> {
-  public run() {
-    const { params, body } = this.node;
+    public run() {
+        const { params, body } = this.node;
 
-    // @TODO: Handle other types of params and pass them directly to next node instead of saving them in the global scope
-    const fn = namedFunction('anonymous function', (args: any[]) => {
-      let index = 0;
+        // @TODO: Handle other types of params and pass them directly to next node instead of saving them in the global scope
+        const fn = namedFunction('anonymous function', (args: any[]) => {
+            let index = 0;
 
-      for (const param of params) {
-        this.visitor.visitNode(param);
+            for (const param of params) {
+                this.visitor.visitNode(param);
 
-        if (param.type === 'Identifier') {
-          this.visitor.scope.set(param.name, args[index]);
-        } else {
-          console.warn('Unhandled param type', param.type);
-        }
+                if (param.type === 'Identifier') {
+                    this.visitor.scope.set(param.name, args[index]);
+                } else {
+                    console.warn('Unhandled param type', param.type);
+                }
 
-        index++;
-      }
+                index++;
+            }
 
-      return this.visitor.visitNode(body);
-    });
+            return this.visitor.visitNode(body);
+        });
 
-    return fn;
-  }
+        return fn;
+    }
 }