+++ src/types/Ytdl.ts
@@ -45,21 +45,21 @@ export type YTDL_Chapter = {
 };
 
 export type YTDL_VideoDetailsAdditions = {
-    videoUrl: string;
     ageRestricted: boolean;
     likes: number | null;
     media: YTDL_Media | null;
     author: YTDL_Author | null;
     storyboards: Array<YTDL_Storyboard>;
     chapters: Array<YTDL_Chapter>;
-    thumbnails: Array<YT_Thumbnail>;
-    description: string | null;
 };
 
 export type YTDL_VideoDetails = YTDL_VideoDetailsAdditions & {
+    videoUrl: string;
     videoId: string;
     title: string;
     playabilityStatus: YT_PlayerApiResponse['playabilityStatus']['status'] | 'UNKNOWN';
+    thumbnails: Array<YT_Thumbnail>;
+    description: string | null;
     lengthSeconds: number;
     keywords: Array<string>;
     channelId: string;
@@ -123,23 +123,21 @@ export type YTDL_VideoInfo<F = YTDL_VideoFormat> = {
     formats: Array<F>;
     full: boolean;
     live_chunk_readahead?: number;
-
     _metadata: {
         isMinimumMode: boolean;
         clients: Array<YTDL_ClientTypes>;
-        html5PlayerUrl: string;
+        html5PlayerId: string;
         id: string;
         options: YTDL_GetInfoOptions;
     };
     _ytdl: {
         version: string;
     };
-
     _playerApiResponses?: {
-        webCreator: YT_PlayerApiResponse | null;
-        tvEmbedded: YT_PlayerApiResponse | null;
-        ios: YT_PlayerApiResponse | null;
-        android: YT_PlayerApiResponse | null;
+        webCreator?: YT_PlayerApiResponse | null;
+        tvEmbedded?: YT_PlayerApiResponse | null;
+        ios?: YT_PlayerApiResponse | null;
+        android?: YT_PlayerApiResponse | null;
         web?: YT_PlayerApiResponse | null;
         mweb?: YT_PlayerApiResponse | null;
         tv?: YT_PlayerApiResponse | null;