+++ src/types/youtube/Formats.ts
@@ -2,7 +2,7 @@ export type YT_Itag = number;
 export type YT_Quality = 'tiny' | 'small' | 'medium' | 'large' | 'hd720' | 'hd1080' | 'hd1440' | 'hd2160' | 'highres' | (string & {});
 export type YT_QualityLabel = '144p' | '144p 15fps' | '144p60 HDR' | '240p' | '240p60 HDR' | '270p' | '360p' | '360p60 HDR' | '480p' | '480p60 HDR' | '720p' | '720p60' | '720p60 HDR' | '1080p' | '1080p60' | '1080p60 HDR' | '1440p' | '1440p60' | '1440p60 HDR' | '2160p' | '2160p60' | '2160p60 HDR' | '4320p' | '4320p60' | 'video' | 'audio';
 
-export type YT_StreamingAdaptiveFormat = {
+export type YT_StreamingAdaptiveFormat<D = false> = {
     itag: YT_Itag;
     url: string;
     signatureCipher?: string;
@@ -33,4 +33,6 @@ export type YT_StreamingAdaptiveFormat = {
     maxDvrDurationSec?: number;
     audioSampleRate?: string;
     audioChannels?: number;
+
+    _deciphered?: D;
 };