+++ src/utils/UserAgents.ts
@@ -58,7 +58,7 @@ const USER_AGENTS: Record<UserAgentType, Array<string>> = {
 class UserAgent {
     static default = 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/127.0.0.0 Safari/537.36';
     static ios = 'com.google.ios.youtube/19.29.1 (iPhone16,2; U; CPU iOS 17_5_1 like Mac OS X;)';
-    static android = 'com.google.android.youtube/18.48.37(Linux; U; Android 13; en_US; sdk_gphone64_x86_64 Build/UPB4.230623.005) gzip';
+    static android = 'com.google.android.youtube/19.35.36(Linux; U; Android 13; en_US; SM-S908E Build/TP1A.220624.014) gzip';
     static tv = 'Mozilla/5.0 (ChromiumStylePlatform) Cobalt/Version';
 
     static getRandomUserAgent(type: UserAgentType): string {