+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## v6.0.6 (2024/11/02)
+
+### Features
+* **URL Decipherment:** Added HTML5 player related options (see below for details).
+```ts
+new YtdlCore({
+    html5Player: {
+        /* It is taken from the html5 player fallback on Github and is automatically updated 6 times a day.
+        * Note: If this option is enabled, Base.js is not retrieved, only the function is retrieved.
+        * */
+        useRetrievedFunctionsFromGithub: true,
+    }
+});
+```
+* **VideoFormat:** Added static function `toVideoFormats` to YtdlCore instance to convert `YT_StreamingAdaptiveFormat` to `YTDL_VideoFormat`.
+
+### Changes
+* **URL Decipherment:** Change to extract the decoding function and the N conversion function when the HTML5 player is acquired.
+
 ## v6.0.5 (2024/10/28)
 
 ### Features
@@ -7,7 +26,7 @@
 * **OAuth2:** Add static function `createOAuth2Credentials` to YtdlCore instances.
 * **Request:** Add option `disableRetryRequest`.
 
-### Change
+### Changes
 * **Download URL:** Changed to output to debug log which URL is being requested when downloading
 * **Default Client:** Changed default client from `['web', 'webCreator', 'tvEmbedded', 'ios', 'android']` to `['web', 'mweb', 'tv', 'ios']`.
 * **Request:** Changed to request again without specifying OriginalProxy or OAuth2 if request fails in order to increase possibility of information retrieval.
@@ -90,7 +109,7 @@ ytdl.download<YTDL_DefaultStreamType>('https://www.youtube.com/watch?v=ID', { fi
 });
 ```
 
-### Change
+### Changes
 * **Lang:** Remove the `lang` option and add the `hl` option.
 * **Country:** Add `gl` option.
 * **Debug:** Eliminated the use of the `YTDL_DEBUG` environment variable. (For debug log display, use the `logDisplay` option instead.)